pyproject.toml
[project]
name = "lookit-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
    "bcrypt==3.2.0",
    "boto3==1.20.50",
    "celery==4.4.7",
    "ciso8601==2.1.3",
    "django==3.2.25",
    "django-ace-overlay",
    "django-bitfield==2.1.0",
    "django-bootstrap-icons==0.8.2",
    "django-bootstrap5==22.2",
    "django-celery-beat==2.0.0",
    "django-cors-headers==3.13.0",
    "django-countries==7.2.1",
    "django-extensions==3.1.5",
    "django-filter==2.4.0",
    "django-guardian==2.3.0",
    "django-localflavor==3.1",
    "django-model-utils==4.0.0",
    "django-multiselectfield==0.1.12",
    "django-prettyjson==0.4.1",
    "django-revproxy",
    "django-storages==1.14.2",
    "djangorestframework==3.12.4",
    "djangorestframework-jsonapi==3.1.0",
    "docker==4.2.1",
    "drf-nested-routers==0.93.3",
    "fleep==1.0.1",
    "gevent==24.2.1",
    "google-cloud-storage==2.1.0",
    "greenlet==3.0.3",
    "kombu==4.6.11",
    "lark-parser==0.12.0",
    "more-itertools==8.8.0",
    "pillow==10.3.0",
    "psycogreen==1.0.2",
    "psycopg2-binary==2.9.9",
    "pydenticon==0.3.1",
    "pyotp==2.6.0",
    "python-dateutil==2.8.1",
    "qrcode==6.1",
    "requests==2.32.3",
    "sendgrid-django==4.2.0",
    "sentry-sdk==1.17.0",
    "transitions==0.9.1",
    "uwsgi>=2.0.29",
]

[dependency-groups]
dev = [
    "beautifulsoup4==4.13.3",
    "coverage==7.6.12",
    "django-dynamic-fixture==4.0.1",
    "django-sslserver==0.22",
    "libsass==0.23.0",
    "parameterized==0.9.0",
    "pre-commit==4.1.0",
    "pyinstrument==5.0.1",
    "python-dotenv==1.0.1",
    "ruff==0.9.10",
]

[tool.uv.sources]
django-ace-overlay = { git = "https://github.com/lookit/django-ace-overlay.git", rev = "master" }
django-revproxy = { git = "https://github.com/Innovativity/django-revproxy.git", rev = "b9fa8375d03fd68747dcb7273a97c19d788aa51b" }

[tool.ruff]
lint.select = ["F", "I", "TID"]

[tool.setuptools.packages.find]
include = ["accounts", "api", "exp", "project", "studies", "web"]
